fix compiler error of determinize
