+++ k2/csrc/determinize.h
@@ -1,0 +1,686 @@
+
+// k2/csrc/determinize.cc
+
+// Copyright (c)  2020  Xiaomi Corporation (authors: Daniel Povey
+// dpove@gmail.com, Haowen Qiu qindazhu@gmail.com)
+
+// See ../../LICENSE for clarification regarding multiple authors
+
+#ifndef K2_CSRC_DETERMINIZE_H_
+#define K2_CSRC_DETERMINIZE_H_
+
+#include <algorithm>
+#include <memory>
+#include <queue>
+#include <unordered_map>
+#include <unordered_set>
+#include <utility>
+#include <vector>
+
+#include "k2/csrc/fsa.h"
+#include "k2/csrc/util.h"
+#include "k2/csrc/weights.h"
+
+namespace k2 {
+/*
+  HOW THIS WORKS
+
+  This is FSA determinization that also outputs derivative information that says
+  how the weights on the arcs in the output FSA vary with the weights on the
+  arcs in the input FSA.
+
+  INTRO TO DETERMINIZATION.
+
+  The problem in determinization of a weighted FSA is to find a deterministic
+  FSA (i.e. one that has no two arcs leaving any given state with the same
+  symbol on), which is equivalent to the input FSA (meaning: the weight it
+  assigns to any given symbol-sequence is the same as the input FSA).  In this
+  explanation, assume epsilons don't exist, if the input FSA had epsilons we'd
+  get rid of them prior to determinization.
+
+
+  SUBSET-BASED ALGORITHMS
+
+   In general, in determinization algorithms, states in the output FSA
+  correspond to weighted subsets of states in the input FSA.  The overall
+  structure of these algorithms will be:
+
+   - Let state 0 in the output FSA correspond to the weighted subset { 0, 0.0 }
+     in the input FSA, where the 0 is the start state-id in the input FSA
+     and the 0.0 is the weight (interpret this as a log-prob).
+     Put that in the queue.
+
+   - While (queue is not empty)
+     Pop output-state from the queue
+     Process successor states of this output-state.
+
+  Obviously most of the detail in the outline above resides in "Process
+  successor states of this output-state."  Let's discuss the unweighted
+  case first.
+
+   *Unweighted case
+
+   Each output-state corresponds to some subset of input-states, say, { s1, s2,
+   .. }.  The set of labels on arcs leaving the output-state will correspond to
+   the set of labels on all the arcs leaving s1, s2 and so on; and the
+   destination-states of those arcs will correspond to the sets of
+   destination-states of those arcs.  The algorithm requires us to store a map
+   from (subset of input-state-ids) to (output-state-id).
+
+   *Weighted case
+
+   In the weighted case, the difference is that instead of a set of
+   input-state-ids we have a weighted subset of them, and the map is from these
+   weighted subsets to state-ids in the output FSA.  The weights in the weighted
+   subsets have to be normalized somehow.  The natural normalization is in the
+   "max/tropical-semiring" case to have the most negative weight be 0.0, and in
+   the "log-sum/log-semiring" case to have the log-sum be 0.0.  Imagine
+   we have a function:
+      Normalize (unnormalized-weighted-subset) -> normalized-weighted-subset,
+  leftover-weight E.g., in the Max case: Normalize( { (6, 1.0), (7, 5.0) } ) ->
+  { (6, 0.0), (7, 4.0) }, 1.0 The "leftover-weights" become the weights on the
+  arcs in the output FSA.
+
+
+   *The problem with differentability
+
+   Consider how to differentiate the weights of the output weighted FSA
+   w.r.t. those of the input.  The problem with differentiability if we use the
+   algorithm above is the case of special symmetries.  What if two weighted
+   subsets happen to coincide because there was an exact relationship between
+  the values of the weights in the input FSA, but there was no *structural*
+  reason in the input FSA why those weighted subsets have to be the same?  Then
+  we have a problem with how to differentiate, because any small change in the
+   input weights would lead to a structural change in the output FSA.
+
+
+  OUR ALGORITHM
+
+    *Different representation of subsets
+
+    Our algorithm is still very similar to the subset-based algorithms mentioned
+    above, and it still involves weighted subsets, but we use a different
+    representation of them.  Our representation (think of this as the key in
+    the map) is:  ( base_state, symbol_sequence ).  The relationship with
+    the weighted subset is: start from state `base_state` in the input FSA,
+    traverse all sequences of arcs that have sequence `symbol_sequence` on them,
+    and the weighted set of states you end up with is the weighted subset
+    in the algorithm above.
+
+    *Different normalization
+
+    Our form of "normalization" of this representation is differen too.  The
+    normalization is to make `symbol_sequence` as short as possible, and advance
+    `base_state` to compensate.  For instance, if `symbol_sequence` is `a b c
+    d`, but the weighted subset of states we can reach by this symbol sequence
+    is the same as what we'd get by moving `base_state` forward two steps and
+    letting `symbol_sequence` be just `c d`, then the latter representation is
+    the canonical one (assuming that was the largest prefix we could remove).
+    Algorithmically, finding the "most recent base_state" involves finding the
+    most recent common ancestor in a directed tree of paths through the input
+    FSA (in the max case) or a directed lattice of paths through the input FSA
+    (in the log-sum case).
+
+    The weights on arcs are related to the total weight of paths from `original
+    base_state` to `new base_state`, (counting only paths that have the removed
+    symbol sequence `a b`).  Just as with the subset algorithm, these
+    weights are what gets "spit out" by the normalization process; we simply
+    have a different normalization process.
+
+
+  PRUNED DETERMINIZATION
+
+    We support pruned determinization.  We won't describe all of the details
+    here, but it involves a priority queue of determinized states, so we always
+    process the "best" queue element first, and we may terminate before the
+    queue is empty.
+
+
+  IMPLEMENTATION DETAILS
+
+    A few details on the implementation:
+
+     - To save memory space, the process of hashing from `base_state,
+  symbol_seq` to output state-id maps them to a fixed-size 128-bit value.  This
+  could in principle generate collisions which would generate incorrect output,
+       but we consider that vanishingly improbable.
+
+ */
+
+struct MaxTracebackState {
+  using DerivType = int32_t;
+
+  int32_t state_id;  // state-id in the input FSA
+
+  int32_t arc_id;  // arc-id in input FSA of the arc that enters state
+                   // `state_id` (or -1 if this is the start state).  It will
+                   // be the best arc if there were multiple possible arcs
+                   // from the base_state to this state with the same symbol
+                   // sequence.
+
+  // prev_state is the state we trace back to (the previous state), which is the
+  // src_state of the arc numbered arc_id.  It will be nullptr if state_id == 0
+  // (start state).
+  std::shared_ptr<MaxTracebackState> prev_state;
+
+  double forward_prob;  // The best forward log-probability from the start
+                        // state to this state (along whichever specific
+                        // sequence of symbols we took to get here)
+
+  // This constructor is for the start-state (state zero) of the input FSA.
+  MaxTracebackState()
+      : state_id(0), arc_id(-1), prev_state(nullptr), forward_prob(0.0) {}
+
+  /**
+     @param [in] state_id  State in input FSA that this corresponds to
+     @param [in] src   Previous LogSumTracebackState that we'll point back
+                      to, or NULL
+     @param [in] incoming_arc_index  Arc-index in input FSA.
+                      Its src_state will equal src->state_id,
+                      its dest_state will equal state_id.
+     @param [in] arc_weight   Weight on the input arc
+   */
+  MaxTracebackState(int32_t state_id,
+                    const std::shared_ptr<MaxTracebackState> &src,
+                    int32_t incoming_arc_index, int32_t arc_weight)
+      : state_id(state_id),
+        arc_id(incoming_arc_index),
+        prev_state(src),
+        forward_prob(src->forward_prob + arc_weight) {}
+
+  /*
+    This takes the same args as the constructor.  It will update the traceback
+    info if this incoming arc had higher weight.
+  */
+  void Accept(const std::shared_ptr<MaxTracebackState> &src, int32_t arc_index,
+              float arc_weight) {
+    double new_forward_prob = src->forward_prob + arc_weight;
+    if (new_forward_prob > forward_prob) {
+      forward_prob = new_forward_prob;
+      arc_id = arc_index;
+      prev_state = src;
+      // state_id doesn't change.
+    }
+  }
+};
+
+class LogSumTracebackState;
+
+/*
+  This struct is used inside LogSumTracebackState; it represents an
+  arc that traces back to a previous LogSumTracebackState.
+  A LogSumTracebackState represents a weighted colletion of paths
+  terminating in a specific state.
+*/
+struct LogSumTracebackLink {
+  std::shared_ptr<LogSumTracebackState> prev_state;
+  // `prev_state` is the state that this points back to.
+
+  int32_t arc_index;  // Index (in input FSA) of this arc from prev_state to the
+                      // destination state (in whose LogSumTracebackState this
+                      // LogSumTracebackLink will be located).
+
+  double
+      forward_prob;  // The total forward log-probability from the start
+                     // state to the end of this arc just before it joins the
+                     // node (conceptually).  Note: this is only the total
+                     // forward log-prob limited to whatever symbol-sequence
+                     // got us to this point...  there may be other
+                     // symbol-sequences terminating in this state.
+                     // (That symbol-sequence can be obtained by tracing back
+                     // this data structure till you hit a LogSumTracebackState
+                     // with prev_state == nullptr (and state_id == 0).
+
+  LogSumTracebackLink(const std::shared_ptr<LogSumTracebackState> &src,
+                      int32_t arc_index, float arc_weight);
+};
+
+/*
+  This stores traceback information for the log-sum case.  Rather than a tree
+  structure, the LogSumTracebackStates interconnect with a lattice structure.
+
+  It can be thought of as as a weighted set of paths from the start state to a
+  particular state.  It will be limited to the subset of paths that have a
+  specific symbol sequence.
+*/
+struct LogSumTracebackState {
+  using DerivType = std::pair<int32_t, float>;
+
+  // `prev_elements` is, conceptually, a list of incoming arcs with associated
+  // weights.
+  std::vector<LogSumTracebackLink> prev_elements;
+
+  int32_t state_id;  // The state-id in the input FSA that this
+                     // LogSumTracebackState corresponds to.
+
+  double forward_prob;  // The total forward log-probability from the start
+                        // state to this state (along whichever specific
+                        // sequence of symbols we took to get here; it's not
+                        // necessarily the best forward-prob in the lattice
+                        // that would take us to this state).  Will equal the
+                        // log-sum of the forward_probs of the prev_elements.
+
+  double backward_prob;  // Used temporarily in algorithms as a backward prob.
+                         // Undefined most of the time.
+
+  // This constructor is to be used only for the start-state (of both the
+  // input FSA and the determinized FSA).
+  LogSumTracebackState() : state_id(0), forward_prob(0.0) {}
+
+  /*
+     @param [in] state_id  State in input FSA
+     @param [in] src  Previous LogSumTracebackState that we'll point back to
+     @param [in] incoming_arc_index.  Arc-index in input FSA.
+                      Its src_state will equal src->state_id, its dest_state
+                      will equal state_id.
+     @param [in] arc_weight   Weight on the arc
+   */
+  LogSumTracebackState(int32_t state_id,
+                       const std::shared_ptr<LogSumTracebackState> &src,
+                       int32_t incoming_arc_index, int32_t arc_weight)
+      : state_id(state_id), forward_prob(src->forward_prob + arc_weight) {
+    prev_elements.emplace_back(src, incoming_arc_index, forward_prob);
+  }
+
+  /*
+     Accept a new incoming link.  The args are the same as for the constructor
+     just above; see documentation there.
+
+      @param [in] src  Previous LogSumTracebackState that we'll point back to
+      @param [in] incoming_arc_index.  Arc-index in input FSA.
+                      Its src_state will equal src->state_id, its dest_state
+                      will equal this->state_id.
+     @param [in] arc_weight   Weight on the incoming arc
+
+   */
+  void Accept(const std::shared_ptr<LogSumTracebackState> &src,
+              int32_t arc_index, float arc_weight) {
+    double link_forward_prob = src->forward_prob + arc_weight;
+    prev_elements.emplace_back(src, arc_index, link_forward_prob);
+    this->forward_prob = LogAdd(this->forward_prob, link_forward_prob);
+  }
+};
+/*
+  Find the most recent common ancestor LogSumTracebackState of a set of
+  LogSumTracebackStates, and return the number of links we had to follow to get
+  there (i.e. the length of symbol sequence).
+
+    @param [in,out] cur_states   A set of TracebackStates that we'll
+                  trace back from.  Must be nonempty.  Equality
+                  here is simply pointer identity.
+
+                  At exit it will contain a single member which will
+                  be the most recent common ancestor.
+
+    @return  Returns the number of links we had to follow (>=0).  If
+             `cur_states.size() == 1` this will be zero.
+ */
+int32_t GetMostRecentCommonAncestor(
+    std::unordered_set<LogSumTracebackState *> *cur_states);
+
+// Version of GetMostRecentCommonAncestor() for MaxTracebackState;
+// see documentation for the other version.
+int32_t GetMostRecentCommonAncestor(
+    std::unordered_set<MaxTracebackState *> *cur_states);
+
+/**
+   A TraceBack() function exists for LogSumTracebackState and MaxTracebackState;
+   it's used in DetState::Normalize().  It finds the cost and derivative
+   information from getting rid of `num_steps` symbols from a symbol sequence.
+
+       @param [in] cur_states   (This is consumed destructively, i.e. don't
+                       expect it to contain the same set on exit).
+                       A set of states; we'll iteratively trace back this
+                       set one step at a time.    At entry it must have
+                       size() == 1; it will also have size() == 1 at exit.
+       @param [in] num_steps   The number of steps to trace back
+       @param [in] arc_weights_in  Weights on the arcs of the input FSA
+       @param [out] weight_out  The output weight; will be the forward-backward
+                       weight of the sub-graph whose final-state is
+                       (*cur_states).front() and whose start-state is
+                       the result of following that back for `num_steps` steps
+                       (which will also be a single state, by virtue of how
+                       the whole determinization algorithm works).  Will be
+                       zero if num_steps == 0.
+       @param [out] deriv_out  Some derivative information at the output
+                       will be written to here, which tells us how the weight
+                       `weight_out` varies as a function of the weights
+                       on the arcs of the input FSA; it's a list
+                       (input_arc_id, deriv) where, mathematically, 0 < deriv <=
+   1 (but we might still get exact zeros due to limitations of floating point
+   representation). Note: the sum of the float values in this vector at exit
+   should be equal to `num_steps`.
+
+ */
+void TraceBack(std::unordered_set<LogSumTracebackState *> *cur_states,
+               int32_t num_steps, const float *arc_weights_in,
+               float *weight_out,
+               std::vector<std::pair<int32_t, float>> *deriv_out);
+
+// The TraceBack function for MaxTracebackState.  See documentation of TraceBack
+// for LogSumTracebackState, above.  This version is simpler.
+void TraceBack(std::unordered_set<MaxTracebackState *> *cur_states,
+               int32_t num_steps,
+               const float *,  // arc_weights_in, unused.
+               float *weight_out, std::vector<int32_t> *deriv_out);
+
+template <class TracebackState>
+class DetState;
+
+template <class TracebackState>
+struct DetStateCompare {
+  bool operator()(const std::shared_ptr<DetState<TracebackState>> &a,
+                  const std::shared_ptr<DetState<TracebackState>> &b) {
+    return a->forward_backward_prob < b->forward_backward_prob;
+  }
+};
+// Priority queue template arguments:
+//   item queued = unique_ptr<DetState> (using pointer equality as comparison)
+//   container type = vector<unique_ptr<DetState> >
+//   less-than operator = DetStateCompare (which compares the
+//   forward_backward_prob).
+template <class TracebackState>
+using DetStatePriorityQueue =
+    std::priority_queue<std::unique_ptr<DetState<TracebackState>>,
+                        std::vector<std::unique_ptr<DetState<TracebackState>>>,
+                        DetStateCompare<TracebackState>>;
+
+template <class TracebackState>
+class DetStateMap;
+/*
+  This represents a determinized state.  Initially it has normalized == false
+  and it represents an un-normalized determinized state (see intro at the top
+  of this file), or an un-normalized determinized state under construction
+  (we add to it using AcceptIncomingArc()).
+
+  After we call Normalize() on it, it is a normalized determinized-state (this
+  also outputs the weight you need for the incoming arc).
+
+  After that
+
+ */
+
+template <class TracebackState>  // TracebackState == MaxTracebackState or
+                                 // LogSumTracebackState
+class DetState {
+ public:
+  using DerivType = typename TracebackState::DerivType;
+  // DerivType == int32_t for MaxTracbackState, or
+  // pair<int32_t, float> for LogSumTracebackState.
+
+  // Constructor for the initial state of the determinized FSA
+  DetState() : seq_len(0), normalized(true) {
+    // the constructor of TracebackState that takes no args gives us what we
+    // need for the start-state.
+    elements[0] = std::make_shared<TracebackState>();
+  }
+
+  /*
+     Constructor (this is the one that's normally used).
+       @param [in] seq_len  Length of symbol sequence from its
+                           base_state (this is before normalization).
+                           Will be the seq_len of the source det_state plus
+                           one.  This seq_len may end up getting reduced
+                           when Normalize() is called (reducing seq_len
+                           implicitly advances the base_state).
+   */
+  DetState(int32_t seq_len)
+      : seq_len(seq_len),
+        normalized(false) {}  // .. and forward_backward_prob undefined
+
+  /**
+     Process incoming arc to this DetState.  See documentation for
+     constructor of TracebackState (which will be MaxTracebackState or
+     LogSumTracebackState).
+         @param [in] state_id  State-id, in input FSA, into which this arc
+     enters. [Note: a DetState is a weighted subset of state-ids.]
+         @param [in] src  The preceding state (from which the arc leaves).
+                    This will be a member of the `elements` of the "parent"
+                    DetState, i.e. the DetState in whose ProcessArcs() function
+                    this DetState was created.
+         @param [in] incoming_arc_index  Arc in input FSA that enters state
+                   `state_id`.
+         @param [in] arc_weight  The weight on this arc
+   */
+  void AcceptIncomingArc(int32_t state_id,
+                         const std::shared_ptr<TracebackState> &src,
+                         int32_t incoming_arc_index, int32_t arc_weight) {
+    auto ret = elements.insert({state_id, nullptr});
+    if (!ret.second) {  // No such state existed in `elements`
+      ret.first->second = std::make_shared<TracebackState>(
+          state_id, src, incoming_arc_index, arc_weight);
+    } else {  // A state with this staste_id existed in `elements`.
+      ret.first->second->Accept(src, incoming_arc_index, arc_weight);
+    }
+  }
+
+  // Length of sequence of symbols (from the base state) leading to this
+  // DetState. each DetState can be described by: start from base_state, follow
+  // paths with a specific symbol sequence, and the weighted set of states that
+  // you reach corresponds to this DetState.
+  //
+  // The sequence of symbols, and the base_state, can be found by tracing back
+  // one of the DetStateElements in the doubly linked list (it doesn't matter
+  // which you pick, the result will be the same).
+  int32_t seq_len;
+
+  // `normalized` is true if this DetState is known to be normalized (meaning:
+  // we have reduced seq_len as much as possible).  DetStates that are not
+  // normalized will not yet have an `output_state`.
+  bool normalized;
+
+  // `elements` can be thought of as weighted subsets of states in the input
+  // FSA, that also stores some traceback information that lets us compute
+  // derivatives.
+  // It's a map from (state-id in input FSA) -> its corresponding
+  // TracebackState.
+  std::unordered_map<int32_t, std::shared_ptr<TracebackState>> elements;
+
+  // This is the weight on the best path that includes this determinized state.
+  // It's needed to form a priority queue on DetStates, so we can process them
+  // best-first.  It is computed as: the forward-weight on `base_state`,
+  // plus the best/most-positive of: (the weight in a DetStateElement plus
+  // the backward-weight of the state associated with that DetStateElement).
+  double forward_backward_prob;
+
+  /*
+    Process arcs leaving this determinized state, possibly creating new
+    determinized states in the process.  Note: Normalize() should already have
+    been called on *this.
+
+              @param [in] wfsa_in  The input FSA that we are determinizing,
+    along with forward-backward weights. The input FSA should normally be
+    epsilon-free as epsilons are treated as a normal symbol; and require
+                                 wfsa_in.weight_tpe == kMaxWeight, for
+                                 now (might later create a version of this code
+                                 that works
+              @param [in] prune_cutoff   Cutoff on forward-backward likelihood
+                                 that we use for pruning; will equal
+                                 wfsa_in.backward_state_weights[0] - prune_beam.
+                                 Will be -infinity if we're not doing pruning.
+              @param [out] arcs_out   Output-FSA arcs-- those leaving this
+                                 determinized state-- will be appended to here.
+              @param [out] arc_weights_out  Weights for the output arcs will
+                                 be appended to here.
+              @param [out] derivs_per_arc  Derivative information for the output
+                                 arcs will be appended to here: either sequences
+                                 of int32_t (corresponding to input arcs), or
+                                 lists of pair<int32_t, float>, corresponding
+                                 to weighted input arcs.
+              @param [in,out] state_map  Maps from DetState to int32_t state-id
+                                 in the output FSA.
+              @return   Returns a number that approximately indicates how much
+                       computation was done (so we can avoid it taking too
+    long).
+  */
+  int32_t ProcessArcs(const WfsaWithFbWeights &wfsa_in, double prune_cutoff,
+                      std::vector<Arc> *arcs_out,
+                      std::vector<float> *arc_weights_out,
+                      std::vector<std::vector<DerivType>> *derivs_per_arc,
+                      DetStateMap<TracebackState> *state_map,
+                      DetStatePriorityQueue<TracebackState> *queue);
+
+  // Computes the forward-backward weight of this DetState.  This is
+  // related to the best cost of any path through the output FSA
+  // that included this determinized state.  I say "related to"
+  // because while it should be exact in the Max case, in the
+  // LogSum case the relationship is a bit more complicated;
+  // maybe just best to say that this is a weight that we use
+  // for pruning.
+  //   @param [in] backward_state_weight   Array, indexed by
+  //                    state in input WFSA, of the weight from this state
+  //                    to the end.  (Of the best path or the sum of paths,
+  //                    depending how it was computed; this will of
+  //                    course affect the pruning).
+  void ComputeFbWeight(const float *backward_state_weights);
+
+  /*
+    Normalizes this DetState by reducing seq_len to the extent possible
+    and outputting the weight and derivative info corresponding to this
+    reduction of sequence length.  Recall that these determinized states
+    are represented as a (base state, and a sequence of symbols that we
+    followed from the base state).  This allows a smaller set of
+    input FSAs to be determinized than the normal weighted-subet-of-states
+    formulation, equivalent (I believe) to the assumption that all
+    the weights are distinct and have no 'special relationships' between
+    them, i.e. no equalities like a + b = c.  This kind of requirement is
+    necessarly for differentiablity.
+
+    This function also sets the forward_backward_prob field.
+
+       @param [in] wfsa_in  The weighted FSA we are determinizing
+       @param [out] removed_weight  The part of the weight that was
+                      removed when we reduced `seq_len`, if any, will
+                      be written to here (else 0.0).
+   */
+  void Normalize(const WfsaWithFbWeights &wfsa_in, float *removed_weight,
+                 std::vector<DerivType> *deriv_info);
+};
+
+/*
+  This class maps from determinized states (DetState) to integer state-ids
+  in the determinized output.  Caution: it uses a randomized algorithm that
+  could in principle produce collisions that would generate wrong output.
+  We don't think this will ever happen though (128-bit representations).
+ */
+template <class TracebackState>
+class DetStateMap {
+ public:
+  /*
+    Looks up the output state-id corresponding to a specific DetState structure,
+    creating a new output-state if necessary.  This does not store any pointers
+    to the DetState or its contents, so you can delete the DetState without
+    affecting this object's ability to map an equivalent DetState to the same
+    state-id.
+
+       @param [in,out] a  The DetState whose state-id we are looking up.
+                         The integer id of the output-FSA state will be written
+                         to its `output_state` field, which at entry is assumed
+                         to be unset.
+        @return  Returns true if this was a NEWLY CREATED state,
+              false otherwise.
+   */
+  bool GetOutputState(DetState<TracebackState> *a) {
+    std::pair<uint64_t, uint64_t> compact;
+    DetStateToCompact(a, &compact);
+    auto p = map_.insert({compact, cur_output_state_});
+    bool inserted = p.second;
+    if (inserted) {
+      a->state_id = cur_output_state_++;
+      return true;
+    } else {
+      a->state_id = p.first->second;
+      return false;
+    }
+  }
+
+  int32_t size() const { return cur_output_state_; }
+
+ private:
+  // simple hashing function that just takes the first element of the pair.
+  struct PairHasher {
+    size_t operator()(const std::pair<uint64_t, uint64_t> &p) const {
+      return static_cast<size_t>(p.first);
+    }
+  };
+
+  int32_t cur_output_state_{0};
+  /* maps from 128-bit key (stored as a pair of uint64_t's) to the int32_t
+   * state-id. */
+  std::unordered_map<std::pair<uint64_t, uint64_t>, uint32_t, PairHasher> map_;
+
+  /* Turns DetState into a compact form of 128 bits.  Technically there
+     could be collisions, which would be fatal for the algorithm, but this
+     is one of those lifetime-of-the-universe type of things (kind of like
+     the theoretical potential for git hash collision) that we ignore.
+
+     The normalized form
+
+  */
+  void DetStateToCompact(const DetState<TracebackState> &d, const Fsa &fsa,
+                         std::pair<uint64_t, uint64_t> *vec) {
+    assert(d.normalized);
+
+    uint64_t a = d.base_state + 17489 * d.seq_len,
+             b = d.base_state * 103979 + d.seq_len;
+
+    // We choose an arbitrary DetStateElement (the first one in the list) to
+    // read the symbol sequence from; the symbol sequence will be the same no
+    // matter which element we choose to trace back.
+    auto elem = d.elements.begin()->second;
+    int32_t seq_len = d.seq_len;
+    const auto &arcs = fsa.arcs;
+    for (int32_t i = 0; i < seq_len; ++i) {
+      int32_t symbol = arcs[elem->arc_id].label;
+      a = symbol + 102299 * a;
+      b = symbol + 102983 * b;
+      elem = elem->prev_state;
+    }
+    vec->first = a;
+    vec->second = b;
+  }
+
+  struct DetStateHasher {
+    size_t operator()(const std::pair<uint64_t, uint64_t> &p) const {
+      return p.first;
+    }
+  };
+};
+
+/*
+  Convenience function that normalizes the state and outputs the arc for it.
+
+  Returns true if the state was newly added (not already present in
+  `state_map`).
+ */
+template <class TracebackState>
+bool NormalizeStateAndOutputArc(
+    DetState<TracebackState> *state, const WfsaWithFbWeights &wfsa_in,
+    float prune_cutoff, std::vector<Arc> *arcs_out,
+    std::vector<float> *arc_weights_out,
+    std::vector<std::vector<typename TracebackState::DerivType>>
+        *derivs_per_arc,
+    DetStateMap<TracebackState> *state_map);
+
+template <class TracebackState>
+double LogSumOrMax(double, double);
+
+template <>
+double LogSumOrMax<MaxTracebackState>(double a, double b) {
+  return std::max(a, b);
+}
+template <>
+double LogSumOrMax<LogSumTracebackState>(double a, double b) {
+  return LogAdd(a, b);
+}
+
+template <typename TracebackState>
+float DeterminizePrunedTpl(
+    const WfsaWithFbWeights &wfsa_in, float beam, int64_t max_step,
+    Fsa *fsa_out, std::vector<float> *arc_weights_out,
+    std::vector<std::vector<typename TracebackState::DerivType>>
+        *arc_derivs_out);
+
+}  // namespace k2
+
+#endif  // K2_CSRC_DETERMINIZE_H_
